wip refactoring into separate crates
