+++ Cargo.toml
@@ -76,10 +76,12 @@ members = [
     "fog/report/api",
     "fog/report/cli",
     "fog/report/connection",
+    "fog/report/resolver",
     "fog/report/server",
     "fog/report/types",
     "fog/report/validation",
     "fog/report/validation/test-utils",
+    "fog/report/verifier",
     "fog/sample-paykit",
     "fog/sig",
     "fog/sig/authority",
