+++ fog/report/resolver/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "mc-fog-report-validation"
+name = "mc-fog-report-resolver"
 version = "2.0.0"
 authors = ["MobileCoin"]
 edition = "2021"
@@ -10,8 +10,12 @@ automock = ["mockall"]
 
 [dependencies]
 mc-account-keys = { path = "../../../account-keys" }
+mc-attest-core = { path = "../../../attest/core", default-features = false }
+mc-attest-verifier = { path = "../../../attest/verifier", default-features = false }
 mc-crypto-keys = { path = "../../../crypto/keys" }
 mc-fog-report-types = { path = "../types" }
+mc-fog-report-validation = { path = "../validation" }
+mc-fog-report-verifier = { path = "../verifier" }
 mc-fog-sig = { path = "../../sig", default-features = false }
 mc-util-encodings = { path = "../../../util/encodings" }
 mc-util-serial = { path = "../../../util/serial" }
