+++ fog/report/verifier/Cargo.toml
@@ -1,22 +1,17 @@
 [package]
-name = "mc-fog-report-validation"
+name = "mc-fog-report-verifier"
 version = "2.0.0"
 authors = ["MobileCoin"]
 edition = "2021"
 
 [features]
 default = []
-automock = ["mockall"]
 
 [dependencies]
-mc-account-keys = { path = "../../../account-keys" }
+mc-attest-core = { path = "../../../attest/core", default-features = false }
+mc-attest-verifier = { path = "../../../attest/verifier", default-features = false }
 mc-crypto-keys = { path = "../../../crypto/keys" }
-mc-fog-report-types = { path = "../types" }
-mc-fog-sig = { path = "../../sig", default-features = false }
 mc-util-encodings = { path = "../../../util/encodings" }
-mc-util-serial = { path = "../../../util/serial" }
-mc-util-uri = { path = "../../../util/uri" }
 
 displaydoc = { version = "0.2", default-features = false }
-mockall = { version = "0.11.2", optional = true }
 serde = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
