+++ mobilecoind/Cargo.toml
@@ -31,6 +31,7 @@ mc-crypto-keys = { path = "../crypto/keys" }
 mc-crypto-rand = { path = "../crypto/rand" }
 mc-crypto-ring-signature-signer = { path = "../crypto/ring-signature/signer" }
 mc-fog-report-connection = { path = "../fog/report/connection" }
+mc-fog-report-resolver = { path = "../fog/report/resolver" }
 mc-fog-report-validation = { path = "../fog/report/validation" }
 mc-ledger-db = { path = "../ledger/db" }
 mc-ledger-migration = { path = "../ledger/migration" }
